AIproject/ai_web_crawler/AIproject/ai_web_crawler/requirements.txt
# 项目依赖库列表
playwright
beautifulsoup4
pyyaml
pydantic
requests
